Move http entities in own package
