+++ src/main/java/de/skillkiller/documentdbackend/controller/AuthorizationController.java
@@ -3,8 +3,8 @@ package de.skillkiller.documentdbackend.controller;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import de.skillkiller.documentdbackend.entity.User;
 import de.skillkiller.documentdbackend.entity.UserDetailsHolder;
-import de.skillkiller.documentdbackend.entity.http.AuthenticationRequest;
-import de.skillkiller.documentdbackend.entity.http.AuthenticationWithTokenRequest;
+import de.skillkiller.documentdbackend.entity.frontend.http.request.AuthenticationRequest;
+import de.skillkiller.documentdbackend.entity.frontend.http.request.AuthenticationWithTokenRequest;
 import de.skillkiller.documentdbackend.search.UserSearch;
 import de.skillkiller.documentdbackend.util.JWTUtil;
 import org.slf4j.Logger;
