+++ src/main/java/de/skillkiller/documentdbackend/controller/DocumentController.java
@@ -4,10 +4,10 @@ import de.skillkiller.documentdbackend.entity.AccessToken;
 import de.skillkiller.documentdbackend.entity.Document;
 import de.skillkiller.documentdbackend.entity.User;
 import de.skillkiller.documentdbackend.entity.UserDetailsHolder;
-import de.skillkiller.documentdbackend.entity.http.SearchResponse;
-import de.skillkiller.documentdbackend.entity.http.UpdateDocumentResponse;
+import de.skillkiller.documentdbackend.entity.frontend.http.response.UpdateDocumentResponse;
 import de.skillkiller.documentdbackend.search.DocumentSearch;
 import de.skillkiller.documentdbackend.search.UserSearch;
+import de.skillkiller.documentdbackend.search.response.SearchResponse;
 import de.skillkiller.documentdbackend.service.AccessTokenService;
 import de.skillkiller.documentdbackend.task.PDFOCR;
 import de.skillkiller.documentdbackend.util.FileUtil;
