+++ src/main/java/de/skillkiller/documentdbackend/entity/frontend/http/request/AuthenticationWithTokenRequest.java
@@ -1,4 +1,4 @@
-package de.skillkiller.documentdbackend.entity.http;
+package de.skillkiller.documentdbackend.entity.frontend.http.request;
 
 import lombok.Data;
