+++ src/main/java/de/skillkiller/documentdbackend/entity/frontend/http/request/PasswordChangeRequest.java
@@ -1,4 +1,4 @@
-package de.skillkiller.documentdbackend.entity.http;
+package de.skillkiller.documentdbackend.entity.frontend.http.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
